requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.2.0
pydantic[email]==2.9.2
PyMuPDF==1.24.9        
pdfplumber==0.11.0    
python-dateutil==2.9.0
autopep8==2.3.1